Mark connection unhandled so bouncy handles git's auth response.
When pushing to the git repo using basic auth the connection is kept
open for git's second auth request with the authentication header.
However, bouncy has marked this connection as handled and returns
without forwarding to ploy's callback.  This fix simply marks the bouncy
connection as unhandled so the second auth request will reach the ploy
callback.

Fixes #3.